python/requirements.txt
python-docx==0.8.11
docx2pdf==0.1.8
python-dotenv==1.0.0
